gameGenWorker/requirements.txt
appwrite
Stockfish
chess
python-dotenv
